remove feature added in last commit because it's behaviour was bug on lang change.
basically when lang changed the number of available translation in the table was not updated and the first language selected was the master in term of number of translation.

this would have removed any good point for this feature subsequently as all language had to have the same number of translation in a table it makes more sense that this number appear as data and not just in translation files.